requirements.txt
# Core
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1

# LangChain/LangGraph for agent orchestration
langchain==1.2.4
langchain-core==1.2.7
langchain-google-genai==4.2.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4

# Google Gemini API
google-genai==1.59.0
google-auth==2.47.0

# HTTP/Networking
httpx==0.28.1
requests==2.32.5
anyio==4.12.1

# Serialization
orjson==3.11.5

# Utilities
tiktoken==0.12.0
tenacity==9.1.2
PyYAML==6.0.3
